requirements.txt
clip==0.2.0
fairscale==0.4.13
image_reward==1.5
numpy==1.24.1
PyYAML==6.0.2
timm==0.6.13
torch==2.1.0
torchmetrics==1.6.0
torchvision==0.16.0
tqdm==4.67.1
transformers==4.45.0
matplotlib
scipy
pillow
pandas
torchinfo
vit
thop
diffusers
accelerate
datasets
huggingface_hub
einops
